+++ src/libAtomVM/opcodesswitch.h
@@ -1106,7 +1106,7 @@ static inline ModuleNativeEntryPoint do_return_native(Module *mod, Context *ctx)
         if (maybe_call_native(ctx, module_name, function_name, fun_arity, &return_value)) { \
             PROCESS_MAYBE_TRAP_RETURN_VALUE(return_value);              \
             x_regs[0] = return_value;                                   \
-            if (ctx->heap.root->next) {                                 \
+            if (memory_heap_fragments_need_gc(&ctx->heap)) {                                 \
                 if (UNLIKELY(memory_ensure_free_with_roots(ctx, 0, 1, x_regs, MEMORY_FORCE_SHRINK) != MEMORY_GC_OK)) { \
                     RAISE_ERROR(OUT_OF_MEMORY_ATOM);                    \
                 }                                                       \
@@ -1919,7 +1919,7 @@ schedule_in:
                         term return_value = nif->nif_ptr(ctx, arity, x_regs);
                         PROCESS_MAYBE_TRAP_RETURN_VALUE_RESTORE_PC_INDEX_ARITY(return_value, orig_pc, mod, index, arity);
                         x_regs[0] = return_value;
-                        if (ctx->heap.root->next) {
+                        if (memory_heap_fragments_need_gc(&ctx->heap)) {
                             if (UNLIKELY(memory_ensure_free_with_roots(ctx, 0, 1, x_regs, MEMORY_FORCE_SHRINK) != MEMORY_GC_OK)) {
                                 RAISE_ERROR(OUT_OF_MEMORY_ATOM);
                             }
@@ -2044,7 +2044,7 @@ schedule_in:
                         ctx->cp = ctx->e[n_words];
                         ctx->e += (n_words + 1);
 
-                        if (ctx->heap.root->next) {
+                        if (memory_heap_fragments_need_gc(&ctx->heap)) {
                             if (UNLIKELY(memory_ensure_free_with_roots(ctx, 0, 1, x_regs, MEMORY_FORCE_SHRINK) != MEMORY_GC_OK)) {
                                 RAISE_ERROR(OUT_OF_MEMORY_ATOM);
                             }
@@ -2333,7 +2333,7 @@ schedule_in:
                 ctx->e += n_words + 1;
                 DEBUG_DUMP_STACK(ctx);
                 // Hopefully, we only need x[0]
-                if (ctx->heap.root->next) {
+                if (memory_heap_fragments_need_gc(&ctx->heap)) {
                     if (UNLIKELY(memory_ensure_free_with_roots(ctx, 0, 1, x_regs, MEMORY_FORCE_SHRINK) != MEMORY_GC_OK)) {
                         RAISE_ERROR(OUT_OF_MEMORY_ATOM);
                     }
@@ -2617,11 +2617,18 @@ schedule_in:
 
                 TRACE("is_eq_exact/3, label=%" PRIu32 ", arg1=%" TERM_X_FMT ", arg2=%" TERM_X_FMT "\n", label, arg1, arg2);
 
-                TermCompareResult result = term_compare(arg1, arg2, TermCompareExact, ctx->global);
-                if (result & (TermLessThan | TermGreaterThan)) {
-                    pc = mod->labels[label];
-                } else if (UNLIKELY(result == TermCompareMemoryAllocFail)) {
-                    RAISE_ERROR(OUT_OF_MEMORY_ATOM);
+                bool fast_equal;
+                if (term_exact_eq_fast(arg1, arg2, &fast_equal)) {
+                    if (!fast_equal) {
+                        pc = mod->labels[label];
+                    }
+                } else {
+                    TermCompareResult result = term_compare(arg1, arg2, TermCompareExact, ctx->global);
+                    if (result & (TermLessThan | TermGreaterThan)) {
+                        pc = mod->labels[label];
+                    } else if (UNLIKELY(result == TermCompareMemoryAllocFail)) {
+                        RAISE_ERROR(OUT_OF_MEMORY_ATOM);
+                    }
                 }
 
                 break;
@@ -2637,11 +2644,18 @@ schedule_in:
 
                 TRACE("is_not_eq_exact/3, label=%" PRIu32 ", arg1=%" TERM_X_FMT ", arg2=%" TERM_X_FMT "\n", label, arg1, arg2);
 
-                TermCompareResult result = term_compare(arg1, arg2, TermCompareExact, ctx->global);
-                if (result == TermEquals) {
-                    pc = mod->labels[label];
-                } else if (UNLIKELY(result == TermCompareMemoryAllocFail)) {
-                    RAISE_ERROR(OUT_OF_MEMORY_ATOM);
+                bool fast_equal;
+                if (term_exact_eq_fast(arg1, arg2, &fast_equal)) {
+                    if (fast_equal) {
+                        pc = mod->labels[label];
+                    }
+                } else {
+                    TermCompareResult result = term_compare(arg1, arg2, TermCompareExact, ctx->global);
+                    if (result == TermEquals) {
+                        pc = mod->labels[label];
+                    } else if (UNLIKELY(result == TermCompareMemoryAllocFail)) {
+                        RAISE_ERROR(OUT_OF_MEMORY_ATOM);
+                    }
                 }
 
                 break;
@@ -2865,12 +2879,19 @@ schedule_in:
                     DECODE_LABEL(jmp_label, pc)
 
                     if (!jump_to_address) {
-                        TermCompareResult result = term_compare(
-                            src_value, cmp_value, TermCompareExact, ctx->global);
-                        if (result == TermEquals) {
-                            jump_to_address = mod->labels[jmp_label];
-                        } else if (UNLIKELY(result == TermCompareMemoryAllocFail)) {
-                            RAISE_ERROR(OUT_OF_MEMORY_ATOM);
+                        bool fast_equal;
+                        if (term_exact_eq_fast(src_value, cmp_value, &fast_equal)) {
+                            if (fast_equal) {
+                                jump_to_address = mod->labels[jmp_label];
+                            }
+                        } else {
+                            TermCompareResult result = term_compare(
+                                src_value, cmp_value, TermCompareExact, ctx->global);
+                            if (result == TermEquals) {
+                                jump_to_address = mod->labels[jmp_label];
+                            } else if (UNLIKELY(result == TermCompareMemoryAllocFail)) {
+                                RAISE_ERROR(OUT_OF_MEMORY_ATOM);
+                            }
                         }
                     }
                 }
@@ -3135,7 +3156,7 @@ schedule_in:
                         PROCESS_MAYBE_TRAP_RETURN_VALUE_LAST(return_value);
                         x_regs[0] = return_value;
 
-                        if (ctx->heap.root->next) {
+                        if (memory_heap_fragments_need_gc(&ctx->heap)) {
                             if (UNLIKELY(memory_ensure_free_with_roots(ctx, 0, 1, x_regs, MEMORY_FORCE_SHRINK) != MEMORY_GC_OK)) {
                                 RAISE_ERROR(OUT_OF_MEMORY_ATOM);
                             }